Revert "add forward and docstrings"
