+++ .pre-commit-config.yaml
@@ -16,7 +16,7 @@ repos:
     rev: v3.2.0
     hooks:
       - id: trailing-whitespace
-      # - id: end-of-file-fixer
+      - id: end-of-file-fixer
       - id: check-yaml
       - id: check-added-large-files
   - repo: https://github.com/PyCQA/flake8
@@ -25,7 +25,7 @@ repos:
       - id: flake8
         args:
           [
-            "--ignore=D100, D200, D205, D400, D401, D104, W503, W504,D401, D101, D107, D102, F401, E402",
+            "--ignore=D100, D200, D205, D400, D401, D104, W503, W504,D401, D101, D107, D102, F401",
             "--max-line-length",
             "88",
           ]
@@ -47,4 +47,3 @@ repos:
 # D400: First line should end with a period
 # D401: First line should be in imperative mood
 # W503 and W504 is contradictory with each other
-# E402: module level import not at top of file
