+++ README.md
@@ -1,24 +1,1 @@
 This repo is made as an assignment of the Cinnamon 2023 Bootcamp.
-
-# Instruction
-
-Step 1: Install the requirements packages
-```
-pip install -r requirements.txt
-```
-
-Step 2: Simply follow the following line to get image converted
-
-```
-python main.py -f <directory to input file> -o <folder of output image(s)>
-```
-
-For example, if you want to use the .tiff file in our 'test' folder and export it into 'outputFiles' folder inside 'test', you can copy the following command
-
-```
-python main.py -f test/inputFiles/untitled-1.pdf -o test/outputFiles
-```
-
-# Author
-
-This repo is made by Ha Noi Team, a team consists of 4 people in Cinnamon Bootcamp 2023
