--- test/targetPaths.py
@@ -1,6 +1,0 @@
-from pathlib import Path
-
-PNG_TARGET_PATH = Path(".") / "results" / "pngTest.png"
-HEIC_TARGET_PATH = Path(".") / "results" / "heicTest.png"
-TIFF_TARGET_PATH = Path(".") / "results" / "tiffTest-1.png"
-PDF_TARGET_PATH = Path(".") / "results" / "untitled-1-1.png"
